+++ client_src/js/printerManager/models/octofarm-printer-log-modal.js
@@ -1,0 +1,9 @@
+document
+  .getElementById(`printerLog-${printer._id}`)
+  .addEventListener("click", async (e) => {
+    let connectionLogs = await OctoFarmClient.get(
+      "printers/connectionLogs/" + printer._id
+    );
+    connectionLogs = await connectionLogs.json();
+    PrinterLogs.loadLogs(printer, connectionLogs);
+  });
