+++ client_src/js/printerManager/models/octofarm-printer-manager.modal.js
@@ -1,0 +1,6 @@
+document
+  .getElementById(`printerButton-${printer._id}`)
+  .addEventListener("click", () => {
+    // eslint-disable-next-line no-underscore-dangle
+    PrinterManager.init(printer._id, printerInfo, printerControlList);
+  });
