+++ client_src/js/printerManager/models/octofarm-printer-statistics-modal.js
@@ -1,0 +1,5 @@
+document
+  .getElementById(`printerStatistics-${printer._id}`)
+  .addEventListener("click", async (e) => {
+    await PrinterLogs.loadStatistics(printer._id);
+  });
