+++ client_src/js/sse.client.js
@@ -1,0 +1,38 @@
+const { parse } = require("flatted");
+
+let source = null;
+
+async function asyncParse(str) {
+  try {
+    const info = parse(str);
+    return info;
+  } catch (e) {
+    return false;
+  }
+}
+
+export function createNewEventSource(url) {
+  source = new EventSource(url);
+
+  source.onmessage = async function (e) {
+    if (e.data != null) {
+      const res = await asyncParse(e.data);
+      postMessage(res);
+    }
+  };
+  source.onerror = function () {
+    postMessage(false);
+  };
+  source.onclose = function () {
+    postMessage(false);
+  };
+}
+
+self.addEventListener(
+  "message",
+  function (e) {
+    const url = e.data.url;
+    createNewEventSource(url);
+  },
+  false
+);
