+++ client_src/package.json
@@ -59,7 +59,6 @@
     "jplist-es6": "^1.2.0",
     "jquery": "^3.5.1",
     "lodash": "^4.17.21",
-    "luxon": "^1.26.0",
     "mini-css-extract-plugin": "^1.6.0",
     "noty": "^3.2.0-beta",
     "prettier": "^2.3.0",
