--- views/partials/PrinterManager/farmStatus.ejs
@@ -1,47 +1,0 @@
-<div class="mt-1 align-items-center d-flex justify-content-center">
-    <button id="offlineCount"
-            type="button"
-            class="btn btn-outline-danger m-1"
-            href="#"
-            aria-expanded="false"
-            disabled
-    >
-        Offline: <br><i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
-    </button>
-    <button id="disconCount"
-            type="button"
-            class="btn btn-danger m-1"
-            href="#"
-            aria-expanded="false"
-            disabled
-    >
-        Disconnected: <br><i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
-    </button>
-    <button id="idleCount"
-            type="button"
-            class="btn btn-dark m-1 "
-            href="#"
-            aria-expanded="false"
-            disabled
-    >
-        Idle: <br><i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
-    </button>
-    <button id="activeCount"
-            type="button"
-            class="btn btn-warning m-1 "
-            href="#"
-            aria-expanded="false"
-            disabled
-    >
-        Active: <br><i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
-    </button>
-    <button id="completeCount"
-            type="button"
-            class="btn btn-success m-1 "
-            href="#"
-            aria-expanded="false"
-            disabled
-    >
-        Complete: <br><i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
-    </button>
-</div>
