+++ views/partials/PrinterManager/printerConnectionLog.ejs
@@ -1,6 +1,6 @@
 <div id="printerTickerMessageBox" class="align-items-center d-flex justify-content-center"
-     style="height:15rem; width:100%; overflow-y: scroll; font-family: monospace;">
-    <div class="spinner-border" role="status">
+     style="height:20rem; width:100%; overflow-y: scroll; font-family: monospace;">
+    <div id="printerTickerLoader" class="spinner-border" role="status">
         <span class="sr-only">Loading...</span>
     </div>
 </div>
